fix: Remove theme overload
PrimeVue provides theme without customization.
The theme provided with installed resource,
and primary colour replacements.
But it is too huge and was not utilizing CDN.
Thus, the new version of `main.css` is importing it from CDN,
and overwrite only the rules that primary colour involved.
Also, there is no need of `theme.css`,
as it is the cached version of PrimeVue theme mentioned before.